# apps/api/src/swagger/swagger_ui.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "POP Api",
		"description": "Documentation sur les webservices de l'application Api"
	},
	"host": "localhost:3000",
	"basePath": "/",
...
	"schemes": ["http"],
	"securityDefinitions": {
		"apiKeyAuth": {
			"type": "apiKey",
			"in": "header",
			"name": "X-API-KEY",
			"description": "any description..."
		}
	},
	"consumes": ["application/json"],
	"produces": ["application/json"],
	"paths": {
...
		"/merimee/newId": {
			"get": {
				"tags": ["Mérimée"],
				"description": "Retourne la nouvelle référence",
				"parameters": [
					{
						"name": "prefix",
						"in": "query",
						"type": "string"
					},
					{
						"name": "dpt",
						"in": "query",
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
		"/merimee/{ref}": {
			"put": {
				"tags": ["Mérimée"],
				"description": "Modification de la notice Mérimée",
				"parameters": [
					{
						"name": "ref",
						"in": "path",
						"required": true,
						"type": "string"
					},
					{
						"name": "body",
						"in": "body",
						"schema": {
							"type": "object",
							"properties": {
								"notice": {
									"example": "any"
								},
								"updateMode": {
									"example": "any"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			},
			"get": {
				"tags": ["Mérimée"],
				"description": "Retourne les informations de la notice Mérimée",
				"parameters": [
					{
						"name": "ref",
						"in": "path",
						"required": true,
						"type": "string",
						"description": "Référence de la notice Mérimée"
					}
				],
				"responses": {
					"200": {
						"description": "Récupération des informations avec succés",
						"schema": {
							"$ref": "#/definitions/GetMerimee"
						}
					},
					"404": {
						"description": "Document non trouvé",
						"schema": {
							"type": "object",
							"properties": {
								"success": {
									"type": "boolean",
									"example": false
								},
								"msg": {
									"type": "string",
									"example": "Document introuvable"
								}
							},
							"xml": {
								"name": "main"
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["Mérimée"],
				"description": "Suppression de la notice Mérimée",
				"parameters": [
					{
						"name": "ref",
						"in": "path",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					},
					"404": {
						"description": "Not Found"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
		"/merimee": {
			"post": {
				"tags": ["Mérimée"],
				"description": "Création de la notice Mérimée",
				"parameters": [
					{
						"name": "body",
						"in": "body",
						"schema": {
							"type": "object",
							"properties": {
								"notice": {
									"example": "any"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
...
	},
	"definitions": {
...
		"GetMerimee": {
			"type": "object",
			"properties": {
				"POP_COORDONNEES": {
					"type": "object",
					"properties": {
						"lat": {
							"type": "number",
							"example": 0
						},
						"lon": {
							"type": "number",
							"example": 0
						}
					}
				},
				"POP_COORDINATES_POLYGON": {
					"type": "object",
					"properties": {
						"type": {
							"type": "string",
							"example": "Polygon"
						},
						"coordinates": {
							"type": "array",
							"example": [],
							"items": {}
						}
					}
				},
				"PRODUCTEUR": {
					"type": "string",
					"example": ""
				},
				"BASE": {
					"type": "string",
					"example": ""
				},
				"CONTIENT_IMAGE": {
					"type": "string",
					"example": ""
				},
				"POP_FLAGS": {
					"type": "array",
					"example": [""],
					"items": {
						"type": "string"
					}
				},
				"POP_CONTIENT_GEOLOCALISATION": {
					"type": "string",
					"example": ""
				},
				"POP_DATE": {
					"type": "array",
					"example": [1500],
					"items": {
						"type": "number"
					}
				},
				"POP_IMPORT": {
					"type": "array",
					"example": ["5c0b8dd6d0c49a0e191ccdf4"],
					"items": {
						"type": "string"
					}
				},
				"TOUT": {
					"type": "string",
					"example": ""
				},
				"ACTU": {
					"type": "string",
					"example": ""
				},
				"ADRS": {
					"type": "string",
					"example": ""
				},
				"AFFE": {
					"type": "string",
					"example": ""
				},
				"AIRE": {
					"type": "string",
					"example": ""
				},
				"APPL": {
					"type": "string",
					"example": ""
				},
				"APRO": {
					"type": "array",
					"example": ["Arrêté"],
					"items": {
						"type": "string"
					}
				},
				"ARCHEO": {
					"type": "string",
					"example": ""
				},
				"AUTP": {
					"type": "array",
					"example": [""],
					"items": {
						"type": "string"
					}
				},
				"AUTR": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"CADA": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"CANT": {
					"type": "string",
					"example": ""
				},
				"COLL": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"COM": {
					"type": "string",
					"example": ""
				},
				"COOR": {
					"type": "string",
					"example": ""
				},
				"COORM": {
					"type": "string",
					"example": ""
				},
				"COPY": {
					"type": "array",
					"example": [""],
					"items": {
						"type": "string"
					}
				},
				"COUV": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"DATE": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"DBOR": {
					"type": "string",
					"example": ""
				},
				"DOMN": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"DENO": {
					"type": "array",
					"example": [""],
					"items": {
						"type": "string"
					}
				},
				"DENQ": {
					"type": "string",
					"example": ""
				},
				"DEPL": {
					"type": "string",
					"example": ""
				},
				"DESC": {
					"type": "string",
					"example": ""
				},
				"DIMS": {
					"type": "string",
					"example": ""
				},
				"DMAJ": {
					"type": "string",
					"example": ""
				},
				"DMIS": {
					"type": "string",
					"example": ""
				},
				"DOSS": {
					"type": "string",
					"example": ""
				},
				"DPRO": {
					"type": "string",
					"example": ""
				},
				"DPT": {
					"type": "string",
					"example": ""
				},
				"DPT_LETTRE": {
					"type": "string",
					"example": ""
				},
				"EDIF": {
					"type": "string",
					"example": ""
				},
				"ELEV": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"ENER": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"ESCA": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"ETAG": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"ETAT": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"ETUD": {
					"type": "array",
					"example": [""],
					"items": {
						"type": "string"
					}
				},
				"GENR": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"HIST": {
					"type": "string",
					"example": ""
				},
				"HYDR": {
					"type": "string",
					"example": ""
				},
				"IMPL": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"INSEE": {
					"type": "string",
					"example": ""
				},
				"INTE": {
					"type": "array",
					"example": [""],
					"items": {
						"type": "string"
					}
				},
				"JATT": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"JDAT": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"LINHA": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"LREG": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"LBASE2": {
					"type": "string",
					"example": ""
				},
				"LIEU": {
					"type": "string",
					"example": ""
				},
				"LOCA": {
					"type": "string",
					"example": ""
				},
				"MFICH": {
					"type": "string",
					"example": ""
				},
				"MOSA": {
					"type": "string",
					"example": ""
				},
				"MHPP": {
					"type": "string",
					"example": ""
				},
				"MICR": {
					"type": "string",
					"example": ""
				},
				"MURS": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"NBOR": {
					"type": "string",
					"example": ""
				},
				"NOMS": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"OBS": {
					"type": "string",
					"example": ""
				},
				"PAFF": {
					"type": "string",
					"example": ""
				},
				"PART": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"PARN": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"PDEN": {
					"type": "string",
					"example": ""
				},
				"PERS": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"PLAN": {
					"type": "string",
					"example": ""
				},
				"PLOC": {
					"type": "string",
					"example": ""
				},
				"PPRO": {
					"type": "string",
					"example": ""
				},
				"PREP": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"PROT": {
					"type": "array",
					"example": [""],
					"items": {
						"type": "string"
					}
				},
				"PSTA": {
					"type": "string",
					"example": ""
				},
				"REFE": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"REFP": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"REFO": {
					"type": "array",
					"example": [""],
					"items": {
						"type": "string"
					}
				},
				"REFJOC": {
					"type": "array",
					"example": [""],
					"items": {
						"type": "string"
					}
				},
				"REFMUS": {
					"type": "array",
					"example": [""],
					"items": {
						"type": "string"
					}
				},
				"REG": {
					"type": "string",
					"example": ""
				},
				"REMA": {
					"type": "string",
					"example": ""
				},
				"REMP": {
					"type": "string",
					"example": ""
				},
				"RENV": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"REPR": {
					"type": "string",
					"example": ""
				},
				"RFPA": {
					"type": "string",
					"example": ""
				},
				"SCLD": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"SCLE": {
					"type": "array",
					"example": [""],
					"items": {
						"type": "string"
					}
				},
				"SCLX": {
					"type": "array",
					"example": ["16e s."],
					"items": {
						"type": "string"
					}
				},
				"SITE": {
					"type": "string",
					"example": ""
				},
				"STAT": {
					"type": "string",
					"example": ""
				},
				"TECH": {
					"type": "array",
					"example": ["Vitrail"],
					"items": {
						"type": "string"
					}
				},
				"TICO": {
					"type": "string",
					"example": ""
				},
				"TOIT": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"TYPO": {
					"type": "string",
					"example": ""
				},
				"VERT": {
					"type": "string",
					"example": ""
				},
				"REFIM": {
					"type": "string",
					"example": ""
				},
				"IMG": {
					"type": "array",
					"example": [""],
					"items": {
						"type": "string"
					}
				},
				"VIDEO": {
					"type": "string",
					"example": ""
				},
				"POP_DOSSIER_VERT": {
					"type": "string",
					"example": ""
				},
				"POP_ARRETE_PROTECTION": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"POP_DOSSIER_PROTECTION": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"DOSURL": {
					"type": "string",
					"example": ""
				},
				"DOSURLPDF": {
					"type": "string",
					"example": ""
				},
				"DOSADRS": {
					"type": "string",
					"example": ""
				},
				"LIENS": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"IMAGE": {
					"type": "string",
					"example": ""
				},
				"VISI": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"VOCA": {
					"type": "string",
					"example": ""
				},
				"VOUT": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"WEB": {
					"type": "string",
					"example": ""
				},
				"ZONE": {
					"type": "string",
					"example": ""
				},
				"THEM": {
					"type": "string",
					"example": ""
				},
				"ACMH": {
					"type": "string",
					"example": ""
				},
				"ACURL": {
					"type": "string",
					"example": ""
				},
				"WADRS": {
					"type": "string",
					"example": ""
				},
				"WCOM": {
					"type": "string",
					"example": "Auxon"
				},
				"WRENV": {
					"type": "string",
					"example": ""
				},
				"REFM": {
					"type": "string",
					"example": ""
				},
				"CONTACT": {
					"type": "string",
					"example": ""
				},
				"IDAGR": {
					"type": "string",
					"example": ""
				},
				"LMDP": {
					"type": "string",
					"example": ""
				},
				"PINT": {
					"type": "string",
					"example": ""
				},
				"DLAB": {
					"type": "string",
					"example": ""
				},
				"_id": {
					"type": "string",
					"example": "5b8418e054caee0d1410972b"
				},
				"POP_COORDINATES_POINT": {
					"type": "object",
					"properties": {
						"type": {
							"type": "string",
							"example": "Point"
						},
						"coordinates": {
							"type": "array",
							"example": [],
							"items": {}
						}
					}
				},
				"POP_HAS_LOCATION": {
					"type": "array",
					"example": ["non"],
					"items": {
						"type": "string"
					}
				},
				"DOSURLP": {
					"type": "string",
					"example": ""
				},
				"REF": {
					"type": "string",
					"example": "PA00078021"
				},
				"MEMOIRE": {
					"type": "array",
					"items": {
						"type": "object",
						"properties": {
							"_id": {
								"type": "string",
								"example": ""
							},
							"ref": {
								"type": "string",
								"example": ""
							},
							"name": {
								"type": "string",
								"example": ""
							},
							"url": {
								"type": "string",
								"example": ""
							},
							"copy": {
								"type": "string",
								"example": ""
							}
						}
					}
				},
				"__v": {
					"type": "number",
					"example": 14
				},
				"POP_COMMENTAIRES": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"HISTORIQUE": {
					"type": "array",
					"items": {
						"type": "object",
						"properties": {
							"_id": {
								"type": "string",
								"example": ""
							},
							"nom": {
								"type": "string",
								"example": ""
							},
							"prenom": {
								"type": "string",
								"example": ""
							},
							"email": {
								"type": "string",
								"example": ""
							},
							"date": {
								"type": "string",
								"example": ""
							},
							"updateMode": {
								"type": "string",
								"example": ""
							}
						}
					}
				}
			}
		},
...
	}
}